Map player session ID to orb color
First pass at a stable color-assignment system that works without coordination
between clients about color choice. Might rewrite this later (or abandon the
idea of assigning colors stably at all), but seems to work astonishingly well
as a first pass to just take the first 6 chars of a hexadecimal session ID
and treat it as a hex color.